1.2.0
#1
Update UI to be responsive for Mobile, Tablet and laptop and up.

a-smiggle
